refactor: Move tests next to the code they test, consolidate fakes
Tests live in the module they test — not in androidApp:
- 7 UseCase/ViewModel tests → usecase/commonTest (KMP-compatible)
- 2 Repository tests → data/commonTest (KMP-compatible)
- 12 tests stay in androidApp (Android-specific: UI, Room, FCM, config)

Fakes consolidated — one copy per interface, in the testing module:
- 5 fakes moved to usecase/commonTest/testfakes/
- 3 fakes moved to data/commonTest/testfakes/
- Deleted duplicate TestDispatcherProvider from androidApp
- Deleted androidApp/testcommon/ directory entirely

All tests converted to kotlin.test (KMP-compatible):
- @Before/@After → @BeforeTest/@AfterTest
- org.junit.Assert → kotlin.test assertions
