--- AndroidGarage/androidApp/src/test/java/com/chriscartland/garage/coroutines/TestDispatcherProvider.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2024 Chris Cartland. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.chriscartland.garage.coroutines
-
-import com.chriscartland.garage.domain.coroutines.DispatcherProvider
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.test.TestDispatcher
-
-class TestDispatcherProvider(
-    testDispatcher: TestDispatcher,
-) : DispatcherProvider {
-    override val main: CoroutineDispatcher = testDispatcher
-    override val io: CoroutineDispatcher = testDispatcher
-    override val default: CoroutineDispatcher = testDispatcher
-}
