--- AndroidGarage/androidApp/src/test/java/com/chriscartland/garage/testcommon/FakeAppLoggerRepository.kt
@@ -1,35 +1,0 @@
-/*
- * Copyright 2024 Chris Cartland. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.chriscartland.garage.testcommon
-
-import com.chriscartland.garage.domain.model.AppLogEvent
-import com.chriscartland.garage.domain.repository.AppLoggerRepository
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-
-class FakeAppLoggerRepository : AppLoggerRepository {
-    val loggedKeys = mutableListOf<String>()
-
-    override suspend fun log(key: String) {
-        loggedKeys.add(key)
-    }
-
-    override fun countKey(key: String): Flow<Long> = MutableStateFlow(0L)
-
-    override fun getAll(): Flow<List<AppLogEvent>> = MutableStateFlow(emptyList())
-}
