+++ AndroidGarage/data/src/commonTest/kotlin/com/chriscartland/garage/data/testfakes/FakeServerConfigRepository.kt
@@ -1,4 +1,4 @@
-package com.chriscartland.garage.testcommon
+package com.chriscartland.garage.data.testfakes
 
 import com.chriscartland.garage.domain.model.ServerConfig
 import com.chriscartland.garage.domain.repository.ServerConfigRepository
