+++ AndroidGarage/usecase/src/commonTest/kotlin/com/chriscartland/garage/usecase/testfakes/FakeSnoozeRepository.kt
@@ -1,4 +1,4 @@
-package com.chriscartland.garage.testcommon
+package com.chriscartland.garage.usecase.testfakes
 
 import com.chriscartland.garage.domain.model.SnoozeRequestStatus
 import com.chriscartland.garage.domain.repository.SnoozeRepository
